Relocate code from SignUp component to new CreateUserButton component
